ml/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pymongo
python-dotenv
motor
